Move assigning equipped slots into slots model
